+++ src/main/java/nhom8/javabackend/hotel/hotel/controller/HotelController.java
@@ -1,9 +1,13 @@
 package nhom8.javabackend.hotel.hotel.controller;
 
-import java.util.List;
+import java.util.Optional;
 
 import javax.validation.Valid;
 
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
+import org.springframework.data.domain.Sort;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -13,6 +17,7 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import nhom8.javabackend.hotel.common.responsehandler.ResponseHandler;
@@ -33,9 +38,10 @@ public class HotelController {
 	}
 
 	@GetMapping
-	public Object findAllHotels() {
-		List<HotelDto> hotels = service.findAllDto();
-		return ResponseHandler.getResponse(hotels, HttpStatus.OK);
+	public Object findAllHotels(@RequestParam("p") Optional<Integer> p) {
+		Pageable pageable= PageRequest.of(p.orElse(0), 5,Sort.by("id"));
+		Page<HotelDto> hotels = service.findAllHotel(pageable);
+		return ResponseHandler.getResponse(service.pagingFormat(hotels), HttpStatus.OK);
 
 	}
 
@@ -61,6 +67,9 @@ public class HotelController {
 
 	@DeleteMapping("/delete-hotel/{hotel-id}")
 	public Object deleteHotel(@PathVariable("hotel-id") Long hotelId) {
+		if(!service.isExistedId(hotelId))
+			return ResponseHandler.getResponse("Hotel doesn't exist",HttpStatus.BAD_REQUEST);
+		
 		service.deleteById(hotelId);
 
 		return ResponseHandler.getResponse(HttpStatus.OK);
